requirements.txt
spotipy==2.23.0
google-cloud==0.34.0
